refactor(client): lift AppNavBar up in component hierarchy
