+++ packages/client/src/App.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 
-import { App } from './main';
+import { App } from './App';
 
 describe('<App />', () => {
   it('should render Home page at launch', async () => {
