+++ packages/client/src/main.tsx
@@ -1,27 +1,9 @@
 import './config/index.css';
 
-import { CssBaseline } from '@mui/material';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
-import { routes } from './config';
-import { ThemeContextProvider } from './contexts';
-import { Home, Main } from './pages';
-
-const App = () => {
-  return (
-    <ThemeContextProvider>
-      <CssBaseline />
-      <BrowserRouter>
-        <Routes>
-          <Route path={routes.home} element={<Home />} />
-          <Route path={routes.main} element={<Main />} />
-        </Routes>
-      </BrowserRouter>
-    </ThemeContextProvider>
-  );
-};
+import { App } from './App';
 
 ReactDOM.render(
   <React.StrictMode>
@@ -29,5 +11,3 @@ ReactDOM.render(
   </React.StrictMode>,
   document.getElementById('root'),
 );
-
-export { App };
