+++ packages/client/src/pages/Main/Main.tsx
@@ -1,14 +1,11 @@
 import Container from '@mui/material/Container';
 
-import { AppNavBar, MediaCardsGrid } from '../../components';
+import { MediaCardsGrid } from '../../components';
 
 const Main = () => (
-  <>
-    <AppNavBar />
-    <Container maxWidth='xl'>
-      <MediaCardsGrid />
-    </Container>
-  </>
+  <Container maxWidth='xl'>
+    <MediaCardsGrid />
+  </Container>
 );
 
 export { Main };
